abstractibg into service object
